backend/requirements.txt
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
astor==0.8.1
asttokens==2.2.1
async-timeout==5.0.1
attrs==25.3.0
Babel==2.12.1
backcall==0.2.0
bandit==1.7.7
beautifulsoup4==4.11.2
billiard==4.2.1
black==24.2.0
bleach==6.0.0
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.15.1
cfgv==3.4.0
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.2
contourpy==1.0.7
coverage==7.8.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
Django==5.1.9
django-cors-headers==4.7.0
django-elasticsearch-dsl==8.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.7
elastic-transport==8.17.1
elasticsearch==8.18.1
elasticsearch-dsl==8.18.0
eralchemy2==1.4.1
et_xmlfile==2.0.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.18.0
flake8==7.0.0
flake8-bugbear==24.4.26
flake8-comprehensions==3.14.0
flake8-docstrings==1.7.0
flake8_simplify==0.21.0
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpointer==2.3
jsonschema==3.2.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
mypy_extensions==1.1.0
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
narwhals==1.41.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
nodeenv==1.9.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==2.2.6
openai==1.82.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pbr==6.1.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.1
pluggy==1.6.0
pre-commit==3.6.0
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==20.0.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.19.1
pygraphviz==1.14
PyJWT==2.9.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==5.0.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.0
sentry-sdk==2.29.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.3.2.post1
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.2
stevedore==5.4.1
streamlit==1.45.1
stripe==12.2.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.17.1
threadpoolctl==3.6.0
tinycss2==1.2.1
tkcalendar==1.5.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.2.0
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.6
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
# Machine Learning
scikit-learn>=1.0.2
pandas>=1.3.0
numpy>=1.21.0
joblib>=1.1.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.10.0
